fix: hide default equipment in admin member view
