+++ miniprogram/pages/role/index.js
@@ -1,89 +1,6 @@
 import { PveService, MemberService } from '../../services/api';
 import { formatStones } from '../../utils/format';
-
-const DEFAULT_EQUIPMENT_IDS = [
-  'novice_sword',
-  'apprentice_helm',
-  'apprentice_robe',
-  'lightstep_boots',
-  'spirit_belt',
-  'initiate_bracers',
-  'initiate_orb',
-  'spirit_ring',
-  'oath_token',
-  'wooden_puppet',
-  'initiate_focus',
-  'initiate_treasure'
-];
-
-const DEFAULT_EQUIPMENT_ID_SET = new Set(DEFAULT_EQUIPMENT_IDS);
-
-function sanitizeEquipmentProfile(profile) {
-  if (!profile || typeof profile !== 'object') {
-    return null;
-  }
-  const sourceEquipment = profile.equipment && typeof profile.equipment === 'object' ? profile.equipment : {};
-  const rawSlots = Array.isArray(sourceEquipment.slots) ? sourceEquipment.slots : [];
-  const rawInventory = Array.isArray(sourceEquipment.inventory) ? sourceEquipment.inventory : [];
-  const sanitizedSlots = rawSlots.map((slot) => {
-    if (!slot || typeof slot !== 'object') {
-      return { slot: '', slotLabel: '', item: null };
-    }
-    const rawItem = slot.item && typeof slot.item === 'object' ? slot.item : null;
-    const item = rawItem && rawItem.itemId && !DEFAULT_EQUIPMENT_ID_SET.has(rawItem.itemId)
-      ? { ...rawItem }
-      : null;
-    return { ...slot, item };
-  });
-  const sanitizedInventory = rawInventory
-    .filter((item) => item && typeof item === 'object' && item.itemId && !DEFAULT_EQUIPMENT_ID_SET.has(item.itemId))
-    .map((item) => ({ ...item }));
-  const setCounts = {};
-  sanitizedSlots.forEach((slot) => {
-    const item = slot && slot.item;
-    if (item && item.setId) {
-      const setId = item.setId;
-      setCounts[setId] = (setCounts[setId] || 0) + 1;
-    }
-  });
-  const rawBonus = sourceEquipment.bonus && typeof sourceEquipment.bonus === 'object' ? sourceEquipment.bonus : {};
-  const rawSets = Array.isArray(rawBonus.sets) ? rawBonus.sets : [];
-  const sanitizedSets = rawSets
-    .map((set) => {
-      if (!set || typeof set !== 'object') {
-        return null;
-      }
-      const setId = set.setId;
-      const count = setId ? setCounts[setId] || 0 : 0;
-      if (!count) {
-        return null;
-      }
-      return { ...set, count };
-    })
-    .filter((set) => !!set);
-  const notes = [];
-  sanitizedSlots.forEach((slot) => {
-    const item = slot && slot.item;
-    if (!item || !Array.isArray(item.notes)) {
-      return;
-    }
-    item.notes.forEach((note) => {
-      if (note && !notes.includes(note)) {
-        notes.push(note);
-      }
-    });
-  });
-  const sanitizedEquipment = {
-    ...sourceEquipment,
-    slots: sanitizedSlots,
-    inventory: sanitizedInventory,
-    bonus: {
-      sets: sanitizedSets,
-      notes
-    }
-  };
-  return { ...profile, equipment: sanitizedEquipment };
-}
+import { sanitizeEquipmentProfile } from '../../utils/equipment';
 
 const ALLOCATABLE_KEYS = ['constitution', 'strength', 'spirit', 'root', 'agility', 'insight'];
